backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
boto3==1.35.0
pydantic==2.9.0
pydantic-settings==2.6.0
python-dotenv==1.0.1
httpx==0.28.0
python-multipart==0.0.12
yt-dlp>=2024.12.0
langdetect==1.0.9

# ── Video Forensics Worker ─────────────────────────────────────────────────
mediapipe==0.10.14
opencv-python-headless==4.10.0.84
scikit-image==0.24.0
scipy==1.14.0
Pillow==11.1.0
numpy==1.26.4

# ── Audio Worker ───────────────────────────────────────────────────────────
librosa==0.10.2
soundfile==0.12.1
faster-whisper==1.1.1
